Upload latest Use Case work
Using Visual Paradigm Online to develop a multi-tier Use Case that characterizes the set of workflows by which a Printer Owner/Operator would Manage the Configuration of their CR6 printer(s).  Looking to capture, abstract and aggregate the set of controls required to support those workflows. Hoping to define an optimal set of screens for all operations EXCEPT printing.