Uploaded wrong DGUS project files at 0.7.1
Still having "teething problems", with my workflow for maintaining the dgus-project files here on GitHub.  GitHub does not seem to recognize when some binary files have changed, so the DGUSTool project files may not all be updated, when I rely on the GitHub Desktop app to update things...
Also lost control of which folder on my system contained the latest release files, so have done my best here to correct the project repo, but cannot be certain that the Source.zip file in the Releases section is 100% correct...

Please rely instead on the DWIN_SET.zip and HTMLDocumentation.zip files to download what you use on your system.